# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonBasketOut": {
        "properties": {
          "comparable": {
            "title": "Comparable",
            "type": "boolean"
          },
          "excluded_item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Item Ids",
            "type": "array"
          },
          "item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Item Ids",
            "type": "array"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "savings_vs_most_expensive": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savings Vs Most Expensive"
          },
          "savings_vs_runner_up": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savings Vs Runner Up"
          },
          "stores": {
            "items": {
              "$ref": "#/components/schemas/StoreTotalsOut"
            },
            "title": "Stores",
            "type": "array"
          },
          "winner_store_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Store Id"
          }
        },
        "required": [
          "item_ids",
          "excluded_item_ids",
          "stores",
          "comparable",
          "note"
        ],
        "title": "CommonBasketOut",
        "type": "object"
      },
      "ComparisonItemOut": {
        "properties": {
          "best_cost": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Cost"
          },
          "best_store_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Store Id"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "cells": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CellOut"
            },
            "title": "Cells",
            "type": "object"
          },
          "image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "worst_cost": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worst Cost"
          }
        },
        "required": [
          "item_id",
          "product_id",
          "name",
          "quantity",
          "unit",
          "cells"
        ],
        "title": "ComparisonItemOut",
        "type": "object"
      },
      "ComparisonOut": {
        "properties": {
          "allow_stale": {
            "title": "Allow Stale",
            "type": "boolean"
          },
          "common": {
            "$ref": "#/components/schemas/CommonBasketOut"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageOut"
          },
          "distance_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Method"
          },
          "freshness_days": {
            "title": "Freshness Days",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "home_located": {
            "title": "Home Located",
            "type": "boolean"
          },
          "include_travel": {
            "title": "Include Travel",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComparisonItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "max_stops": {
            "title": "Max Stops",
            "type": "integer"
          },
          "plan": {
            "$ref": "#/components/schemas/EconomicPlanOut"
          },
          "recommendation": {
            "$ref": "#/components/schemas/RecommendationOut"
          },
          "stores": {
            "items": {
              "$ref": "#/components/schemas/ComparisonStoreOut"
            },
            "title": "Stores",
            "type": "array"
          },
          "vehicle_configured": {
            "title": "Vehicle Configured",
            "type": "boolean"
          }
        },
        "required": [
          "generated_at",
          "freshness_days",
          "allow_stale",
          "include_travel",
          "max_stops",
          "home_located",
          "vehicle_configured",
          "stores",
          "items",
          "common",
          "coverage",
          "plan",
          "recommendation"
        ],
        "title": "ComparisonOut",
        "type": "object"
      },
...
      "CoverageOut": {
        "properties": {
          "best_store_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Store Id"
          },
          "max_coverage": {
            "title": "Max Coverage",
            "type": "integer"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "stores": {
            "items": {
              "$ref": "#/components/schemas/StoreTotalsOut"
            },
            "title": "Stores",
            "type": "array"
          }
        },
        "required": [
          "stores",
          "max_coverage",
          "note"
        ],
        "title": "CoverageOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryOut": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "freshness_days": {
            "title": "Freshness Days",
            "type": "integer"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/HistorySeriesOut"
            },
            "title": "Series",
            "type": "array"
          }
        },
        "required": [
          "product_id",
          "product_name",
          "freshness_days",
          "days",
          "series"
        ],
        "title": "HistoryOut",
        "type": "object"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "Ok",
        "type": "object"
      },
...
      "RunCreateIn": {
        "properties": {
          "allow_llm": {
            "default": true,
            "title": "Allow Llm",
            "type": "boolean"
          },
          "list_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Id"
          },
          "product_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 200,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Ids"
          },
          "store_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 40,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Ids"
          }
        },
        "title": "RunCreateIn",
        "type": "object"
      },
      "RunDetailOut": {
        "properties": {
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "done_targets": {
            "title": "Done Targets",
            "type": "integer"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "list_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Id"
          },
          "llm_calls": {
            "title": "Llm Calls",
            "type": "integer"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/RunTargetOut"
            },
            "title": "Targets",
            "type": "array"
          },
          "total_targets": {
            "title": "Total Targets",
            "type": "integer"
          },
          "trigger": {
            "$ref": "#/components/schemas/RunTrigger"
          }
        },
        "required": [
          "id",
          "status",
          "trigger",
          "created_at",
          "cancel_requested",
          "total_targets",
          "done_targets",
          "counts",
          "llm_calls",
          "targets",
          "events",
          "retryable"
        ],
        "title": "RunDetailOut",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "done_targets": {
            "title": "Done Targets",
            "type": "integer"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "list_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Id"
          },
          "llm_calls": {
            "title": "Llm Calls",
            "type": "integer"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "total_targets": {
            "title": "Total Targets",
            "type": "integer"
          },
          "trigger": {
            "$ref": "#/components/schemas/RunTrigger"
          }
        },
        "required": [
          "id",
          "status",
          "trigger",
          "created_at",
          "cancel_requested",
          "total_targets",
          "done_targets",
          "counts",
          "llm_calls"
        ],
        "title": "RunOut",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "success",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "RunTrigger": {
        "enum": [
          "manual",
          "schedule",
          "cli",
          "retry"
        ],
        "title": "RunTrigger",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Comparação de preços de supermercado com cobertura, frescor e custo de deslocamento.",
    "title": "PriceTracker API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/comparison": {
      "get": {
        "description": "Comparison of the list across stores: common basket, coverage, economic plan, recommendation.",
        "operationId": "comparison_api_v1_comparison_get",
        "parameters": [
          {
            "in": "query",
            "name": "list_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "List Id"
            }
          },
          {
            "in": "query",
            "name": "store_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Store Ids"
            }
          },
          {
            "in": "query",
            "name": "allow_stale",
            "required": false,
            "schema": {
              "default": false,
              "title": "Allow Stale",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_travel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include Travel"
            }
          },
          {
            "in": "query",
            "name": "max_stops",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Stops"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Comparison",
        "tags": [
          "comparison"
        ]
      }
    },
...
    "/api/v1/history/products/{product_id}": {
      "get": {
        "operationId": "product_history_api_v1_history_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 180,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Product History",
        "tags": [
          "history"
        ]
      }
    },
...
    "/api/v1/observations/{observation_id}/review": {
      "post": {
        "operationId": "review_observation_api_v1_observations__observation_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": true,
            "schema": {
              "title": "Decision",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Observation",
        "tags": [
          "history"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs Api V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/retry": {
      "post": {
        "operationId": "retry_run_api_v1_runs__run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}